Compress data by only counting zeroes
This strategy allows us to only use 1 short for the number of zeroes
preceding a rating and 1 char for the rating which adds up to just 3
bytes per rating.  Previous RLE implementation required 6 bytes per
rating on average because it required a short and a char for the zeroes
before a rating as well as a short and a char for the rating.