Rewrite kNN buildMatrix() with improved algorithm
As described in blog post (linked in slack), this algorithm is an improvement on the
brute force algorithm. Instead of looking at all pairs of users (i, j), it checks
only users j that have rated movie m that has been rated by user i.

We get a 2x time improvement, instead of the 10x in the blog post.

Also changed all thresholds to inclusive.